requirements.txt
autopep8==1.4.4
backports.csv==1.0.7
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.0.12
django-authtools==1.7.0
django-bootstrap4==1.0.1
django-heroku==0.3.1
django-import-export==1.2.0
django-postgres-copy==2.4.2
django-registration==2.4.1
django-tinymce==2.8.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
jdcal==1.3
odfpy==1.4.0
openpyxl==2.5.1
Pillow==6.1.0
psycopg2==2.8.3
pycodestyle==2.5.0
python-decouple==3.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
soupsieve==1.9.3
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.6
whitenoise==4.1.3
xlrd==1.2.0
xlwt==1.3.0
